examples/cvi-frontend-backend-tools/requirements.txt
requests
openai
fastapi
python-dotenv
uvicorn
daily-python